Address LGTM.com recommendations (#86)
* LGTM.com recommendation: Unused import

* LGTM.com recommendation: Unused local variable

* LGTM.com recommendation: Except block handles 'BaseException'

All exception classes in Python derive from BaseException.
BaseException has three important subclasses, Exception from which
all errors and normal exceptions derive, KeyboardInterrupt which is
raised when the user interrupts the program from the keyboard and
SystemExit which is raised by the sys.exit() function to terminate
the program.

Since KeyboardInterrupt and SystemExit are special they should
not be grouped together with other Exception classes.

Catching BaseException, rather than its subclasses may prevent
proper handling of KeyboardInterrupt or SystemExit. It is easy to
catch BaseException accidentally as it is caught implicitly by an
empty except: statement.